examples: Update examples/docs to reflect pdu blocks moving to pdu module
The digital and FEC modules had a large number of references to the PDU blocks that were moved from gr-blocks to gr-pdu, this updates these changes in example flowgraphs and a few python files. The usage-manual update will be propagated to the wiki so that future exports will remain up to date.

Signed-off-by: Jacob Gilbert <jacob.gilbert@protonmail.com>